2022/06/26 get_next_line modules

+++ src_memory/ft_memchr.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, long unsigned int n)
+#include <libft.h>
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*sptr;
