+++ src_memory/ft_memcmp.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, long unsigned int n)
+#include <libft.h>
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char	*sptr1;
 	unsigned char	*sptr2;
