+++ src_memory/ft_memcpy.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, long unsigned int n)
+#include <libft.h>
+
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned char	*dptr;
 	unsigned char	*sptr;
